+++ yash-builtin/src/fg.rs
@@ -63,8 +63,13 @@
 //!
 //! # Exit status
 //!
-//! The built-in returns the exit status of the resumed job. On error, it
-//! returns a non-zero exit status.
+//! If a resumed job suspends and the current environment is
+//! [interactive](Env::is_interactive), the built-in returns with the
+//! [`Interrupt`] divert, which should make the shell stop the current command
+//! execution and return to the prompt. Otherwise, the built-in returns with the
+//! exit status of the resumed job.
+//!
+//! On error, it returns a non-zero exit status.
 //!
 //! # Portability
 //!
@@ -86,13 +91,16 @@ use crate::common::report_error;
 use crate::common::report_simple_failure;
 use crate::common::syntax::Mode;
 use crate::common::syntax::parse_arguments;
+use std::ops::ControlFlow::Break;
 use yash_env::Env;
 use yash_env::io::Fd;
 #[cfg(doc)]
 use yash_env::job::JobList;
 use yash_env::job::Pid;
+use yash_env::job::ProcessResult;
 use yash_env::job::ProcessState;
 use yash_env::job::id::parse;
+use yash_env::semantics::Divert::Interrupt;
 use yash_env::semantics::ExitStatus;
 use yash_env::semantics::Field;
 use yash_env::signal;
@@ -101,12 +109,12 @@ use yash_env::system::System as _;
 use yash_env::system::SystemEx as _;
 
 /// Waits for the specified job to finish (or suspend again).
-async fn wait_until_halt(env: &mut Env, pid: Pid) -> Result<ProcessState, Errno> {
+async fn wait_until_halt(env: &mut Env, pid: Pid) -> Result<ProcessResult, Errno> {
     loop {
         let (_pid, state) = env.wait_for_subshell(pid).await?;
         match state {
             ProcessState::Running => (),
-            ProcessState::Halted(_) => return Ok(state),
+            ProcessState::Halted(result) => return Ok(result),
         }
     }
 }
@@ -117,7 +125,7 @@ async fn wait_until_halt(env: &mut Env, pid: Pid) -> Result<ProcessState, Errno>
 /// signal to it. It then waits for the job to finish (or suspend again).
 ///
 /// This function panics if there is no job at the specified index.
-async fn resume_job_by_index(env: &mut Env, index: usize) -> Result<ProcessState, ResumeError> {
+async fn resume_job_by_index(env: &mut Env, index: usize) -> Result<ProcessResult, ResumeError> {
     let tty = env.get_tty()?;
 
     let job = &env.jobs[index];
@@ -132,36 +140,45 @@ async fn resume_job_by_index(env: &mut Env, index: usize) -> Result<ProcessState
     env.system.write_all(Fd::STDOUT, line.as_bytes()).await?;
     drop(line);
 
-    let mut state = job.state;
-    if state.is_alive() {
-        // TODO Should we save/restore the terminal state?
+    let result = match job.state {
+        // If the job is already finished, return the result immediately.
+        ProcessState::Halted(
+            result @ (ProcessResult::Exited(_) | ProcessResult::Signaled { .. }),
+        ) => result,
 
-        // Make sure to put the target job in the foreground before sending the
-        // SIGCONT signal, or the job may be immediately re-suspended.
-        env.system.tcsetpgrp_without_block(tty, job.pid)?;
+        // If the job is still alive, resume it.
+        ProcessState::Halted(ProcessResult::Stopped(_)) | ProcessState::Running => {
+            // TODO Should we save/restore the terminal state?
 
-        let pgid = -job.pid;
-        let sigcont = env.system.signal_number_from_name(signal::Name::Cont);
-        let sigcont = sigcont.ok_or(Errno::EINVAL)?;
-        env.system.kill(pgid, Some(sigcont)).await?;
+            // Make sure to put the target job in the foreground before sending the
+            // SIGCONT signal, or the job may be immediately re-suspended.
+            env.system.tcsetpgrp_without_block(tty, job.pid)?;
 
-        // Wait for the job to finish (or suspend again).
-        state = wait_until_halt(env, job.pid).await?;
+            let pgid = -job.pid;
+            let sigcont = env.system.signal_number_from_name(signal::Name::Cont);
+            let sigcont = sigcont.ok_or(Errno::EINVAL)?;
+            env.system.kill(pgid, Some(sigcont)).await?;
 
-        // Move the shell back to the foreground.
-        env.system.tcsetpgrp_with_block(tty, env.main_pgid)?;
-    }
+            // Wait for the job to finish (or suspend again).
+            let result = wait_until_halt(env, job.pid).await?;
+
+            // Move the shell back to the foreground.
+            env.system.tcsetpgrp_with_block(tty, env.main_pgid)?;
+
+            result
+        }
+    };
 
     // Remove the job if it has finished.
-    if !state.is_alive() {
+    if !result.is_stopped() {
         env.jobs.remove(index);
     }
 
-    Ok(state)
+    Ok(result)
 }
 
 /// Resumes the job specified by the operand.
-async fn resume_job_by_id(env: &mut Env, job_id: &str) -> Result<ProcessState, OperandErrorKind> {
+async fn resume_job_by_id(env: &mut Env, job_id: &str) -> Result<ProcessResult, OperandErrorKind> {
     let job_id = parse(job_id)?;
     let index = job_id.find(&env.jobs)?;
     Ok(resume_job_by_index(env, index).await?)
@@ -189,7 +206,11 @@ pub async fn main(env: &mut Env, args: Vec<Field>) -> crate::Result {
     };
 
     match result {
-        Ok(state) => ExitStatus::try_from(state).unwrap().into(),
+        Ok(result) if env.is_interactive() => {
+            let divert = Break(Interrupt(Some(ExitStatus::from(result))));
+            crate::Result::with_exit_status_and_divert(env.exit_status, divert)
+        }
+        Ok(result) => ExitStatus::from(result).into(),
         Err(error) => report_simple_failure(env, &error.to_string()).await,
     }
 }
@@ -203,7 +224,7 @@ mod tests {
     use yash_env::VirtualSystem;
     use yash_env::job::Job;
     use yash_env::job::ProcessResult;
-    use yash_env::option::Option::Monitor;
+    use yash_env::option::Option::{Interactive, Monitor};
     use yash_env::option::State::On;
     use yash_env::subshell::JobControl;
     use yash_env::subshell::Subshell;
@@ -292,7 +313,7 @@ mod tests {
 
             let result = resume_job_by_index(&mut env, index).await.unwrap();
 
-            assert_eq!(result, ProcessState::exited(42));
+            assert_eq!(result, ProcessResult::exited(42));
             let state = state.borrow().processes[&pid].state();
             assert_eq!(state, ProcessState::exited(42));
             // The finished job should be removed from the job list.
@@ -322,7 +343,7 @@ mod tests {
 
             let result = resume_job_by_index(&mut env, index).await.unwrap();
 
-            assert_eq!(result, ProcessState::stopped(SIGSTOP));
+            assert_eq!(result, ProcessResult::Stopped(SIGSTOP));
             let job_state = env.jobs[index].state;
             assert_eq!(job_state, ProcessState::stopped(SIGSTOP));
             let state = state.borrow().processes[&pid].state();
@@ -371,7 +392,7 @@ mod tests {
 
         let result = resume_job_by_index(&mut env, index).now_or_never().unwrap();
 
-        assert_eq!(result, Ok(ProcessState::exited(12)));
+        assert_eq!(result, Ok(ProcessResult::exited(12)));
         // The finished job should be removed from the job list.
         assert_eq!(env.jobs.get(index), None);
 
@@ -521,5 +542,69 @@ mod tests {
         assert_stderr(&system.state, |stderr| {
             assert!(stderr.contains("not found"), "{stderr:?}");
         });
+    }
+
+    #[test]
+    fn main_returns_exit_status_if_job_suspends_if_not_interactive() {
+        in_virtual_system(|mut env, state| async move {
+            stub_tty(&state);
+            env.options.set(Monitor, On);
+            env.exit_status = ExitStatus(42);
+            let subshell = Subshell::new(|env, _| {
+                Box::pin(async move {
+                    suspend(env).await;
+                    suspend(env).await;
+                    unreachable!("child process should not be resumed twice");
+                })
+            })
+            .job_control(JobControl::Foreground);
+            let (pid, subshell_result) = subshell.start_and_wait(&mut env).await.unwrap();
+            assert_eq!(subshell_result, ProcessResult::Stopped(SIGSTOP));
+            let mut job = Job::new(pid);
+            job.job_controlled = true;
+            job.state = subshell_result.into();
+            let index = env.jobs.add(job);
+            env.jobs.set_current_job(index).unwrap();
+
+            let result = main(&mut env, vec![]).await;
+            assert_eq!(
+                result,
+                crate::Result::from(ExitStatus(SIGSTOP.as_raw() + 0x180))
+            );
+        })
+    }
+
+    #[test]
+    fn main_returns_interrupt_if_job_suspends_if_interactive() {
+        in_virtual_system(|mut env, state| async move {
+            stub_tty(&state);
+            env.options.set(Interactive, On);
+            env.options.set(Monitor, On);
+            env.exit_status = ExitStatus(42);
+            let subshell = Subshell::new(|env, _| {
+                Box::pin(async move {
+                    suspend(env).await;
+                    suspend(env).await;
+                    unreachable!("child process should not be resumed twice");
+                })
+            })
+            .job_control(JobControl::Foreground);
+            let (pid, subshell_result) = subshell.start_and_wait(&mut env).await.unwrap();
+            assert_eq!(subshell_result, ProcessResult::Stopped(SIGSTOP));
+            let mut job = Job::new(pid);
+            job.job_controlled = true;
+            job.state = subshell_result.into();
+            let index = env.jobs.add(job);
+            env.jobs.set_current_job(index).unwrap();
+
+            let result = main(&mut env, vec![]).await;
+            assert_eq!(
+                result,
+                crate::Result::with_exit_status_and_divert(
+                    ExitStatus(42),
+                    Break(Interrupt(Some(ExitStatus(SIGSTOP.as_raw() + 0x180))))
+                )
+            );
+        })
     }
 }